+++ driver/src/main/java/oracle/nosql/driver/iam/Utils.java
@@ -39,6 +39,7 @@ import java.util.TimeZone;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.regex.Pattern;
+import java.util.UUID;
 
 import oracle.nosql.driver.Region;
 import oracle.nosql.driver.iam.CertificateSupplier.X509CertificateKeyPair;
@@ -88,6 +89,12 @@ class Utils {
 
     /* 4k bytes */
     private static final int BUF_SIZE = 0x800;
+
+    /* fields in JWT token JSON used to check validity */
+    private static final String[] FIELDS = {
+        "exp", "jwk", "n", "e",
+        SignatureProvider.ResourcePrincipalClaimKeys.COMPARTMENT_ID_CLAIM_KEY,
+        SignatureProvider.ResourcePrincipalClaimKeys.TENANT_ID_CLAIM_KEY};
 
     /* Map used to lookup IAM URI */
     private static final Map<String, String> IAM_URI = new HashMap<>();
@@ -662,6 +669,113 @@ class Utils {
             }
         }
         return null;
+    }
+
+    /*
+     * Parse JSON response that only has one field token.
+     * { "token": "...."}
+     */
+    static String parseTokenResponse(String response) {
+        try {
+            JsonParser parser = createParser(response);
+            if (parser.getCurrentToken() == null) {
+                parser.nextToken();
+            }
+            while (parser.getCurrentToken() != null) {
+                String field = findField(response, parser, "token");
+                if (field != null) {
+                    parser.nextToken();
+                    return parser.getText();
+                }
+            }
+            throw new IllegalStateException(
+                "Unable to find security token in " + response);
+        } catch (IOException ioe) {
+            throw new IllegalStateException(
+                "Error parsing security token " + response +
+                    " " + ioe.getMessage());
+        }
+    }
+
+    /*
+     * Parse security token JSON response get fields expiration time,
+     * modulus and public exponent of JWK, only used for security token
+     * validity check, ignores other fields.
+     *
+     * Response:
+     * {
+     *  "exp" : 1234123,
+     *  "jwk" : {
+     *    "e": "xxxx",
+     *    "n": "xxxx",
+     *    ...
+     *  }
+     *  ...
+     * }
+     */
+    static Map<String, String> parseToken(String token) {
+        if (token == null) {
+            return null;
+        }
+        String[] jwt = splitJWT(token);
+        String claimJson = new String(Base64.getUrlDecoder().decode(jwt[1]),
+            StandardCharsets.UTF_8);
+
+        try {
+            JsonParser parser = createParser(claimJson);
+            Map<String, String> results = new HashMap<>();
+            parse(token, parser, results);
+
+            String jwkString = results.get("jwk");
+            if (jwkString == null) {
+                return results;
+            }
+            parser = createParser(jwkString);
+            parse(token, parser, results);
+
+            return results;
+        } catch (IOException ioe) {
+            throw new IllegalStateException(
+                "Error parsing security token "+ ioe.getMessage());
+        }
+    }
+
+    private static void parse(String json,
+                              JsonParser parser,
+                              Map<String, String> reults)
+        throws IOException {
+
+        if (parser.getCurrentToken() == null) {
+            parser.nextToken();
+        }
+        while (parser.getCurrentToken() != null) {
+            String field = findField(json, parser, FIELDS);
+            if (field != null) {
+                parser.nextToken();
+                reults.put(field, parser.getText());
+            }
+        }
+    }
+
+    private static String[] splitJWT(String jwt) {
+        final int dot1 = jwt.indexOf(".");
+        final int dot2 = jwt.indexOf(".", dot1 + 1);
+        final int dot3 = jwt.indexOf(".", dot2 + 1);
+
+        if (dot1 == -1 || dot2 == -1 || dot3 != -1) {
+            throw new IllegalArgumentException(
+                "Given string is not in the valid access token format");
+        }
+
+        final String[] parts = new String[3];
+        parts[0] = jwt.substring(0, dot1);
+        parts[1] = jwt.substring(dot1 + 1, dot2);
+        parts[2] = jwt.substring(dot2 + 1);
+        return parts;
+    }
+
+    static String generateOpcRequestId() {
+        return UUID.randomUUID().toString().replace("-", "").toUpperCase();
     }
 
     static void logTrace(Logger logger, String message) {